split up the fast integration tests into pre, tests and cleanup scripts
